requirements.txt
anyio==3.5.0
APScheduler==3.9.1
asgiref==3.5.0
asyncio==3.4.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
fastapi==0.77.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
pydantic==1.9.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uvicorn[standard]==0.17.6
httpx~=0.23.0
aiobotocore~=2.3.2
Authlib~=1.0.1
itsdangerous~=2.1.1